refactor(domains): move vendor-visuals to domains/vendors/visuals (#232)
