+++ src/app/(public)/productores/[slug]/page.tsx
@@ -21,7 +21,7 @@ import { JsonLd } from '@/components/seo/JsonLd'
 import { absoluteUrl, buildPageMetadata } from '@/lib/seo'
 import { getServerLocale } from '@/i18n/server'
 import { getCatalogCopy, getLocalizedCertificationCopy } from '@/i18n/catalog-copy'
-import { getVendorHeroImage, getVendorVisualLabel } from '@/lib/vendor-visuals'
+import { getVendorHeroImage, getVendorVisualLabel } from '@/domains/vendors/visuals'
 import { Badge } from '@/components/ui/badge'
 import { StarRating } from '@/components/reviews/StarRating'
 import { auth } from '@/lib/auth'
