+++ src/domains/vendors/visuals.ts
